Use shared start and end icons in admin map
